moved pkg/healthcheck to internal/healthcheck
